[♻️refactor]: api 관련 type들 types/api 디렉토리로 이동
