+++ src/types/api/componentDetail.ts
